Refactor the subnav
